chore(refactor): move bonds components to modules
