+++ src/common/components/Bonds/OverviewWrapper/Overview.styles.ts
@@ -1,0 +1,28 @@
+import styled from 'styled-components'
+
+export const BondsHomeSectionNav = styled.div`
+  padding: 1rem;
+  margin-left: 1.875rem;
+  a {
+    font-weight: normal;
+    font-size: 1.1875rem;
+    text-transform: uppercase;
+    text-decoration: none;
+
+    color: #ffffff;
+    padding: 0.25rem 1.5rem;
+    &.active {
+      color: #87def6;
+    }
+    &:hover {
+      text-decoration: none;
+      color: #87def6;
+    }
+  }
+  @media (max-width: 768px) {
+    padding: 1rem 0;
+    margin: 0;
+    width: initial;
+    overflow-x: scroll;
+  }
+`
