--- src/components/bonds/TransactionsTable.scss
@@ -1,127 +1,0 @@
-@import '../../assets/styles/_colors';
-@import '../../assets/styles/_fonts';
-
-.Table {
-  .Table_filter__fields {
-    display: flex;
-    font-family: $header-title-font;
-    padding: 1.25rem 0.625rem;
-    font-weight: normal;
-    input,
-    select,
-    button,
-    .DateRangePickerInput {
-      margin: 0 0.625rem;
-      font-family: $header-title-font;
-      background: #f0f4f9;
-      color: #636971;
-      border-radius: 3px;
-      svg path {
-        fill: #636971;
-      }
-    }
-    .orderType {
-      max-width: 100px;
-    }
-    .token {
-      max-width: 100px;
-    }
-
-    @media (max-width: 768px) {
-      width: initial;
-      overflow-x: scroll;
-    }
-  }
-
-  .transaction-table-wrapper {
-    @media (max-width: 768px) {
-      overflow-x: scroll;
-      .transaction-table-inner {
-        width: 55rem;
-      }
-    }
-  }
-  .ReactVirtualized__Table__headerRow {
-    font-size: 0.8rem;
-    padding: 0.625rem 1.25rem 0;
-    font-family: 'Roboto', sans-serif;
-    color: white;
-    font-weight: normal;
-  }
-  .ReactVirtualized__Table__headerRow,
-  .ReactVirtualized__Table__Grid,
-  .ReactVirtualized__Grid__innerScrollContainer,
-  .ReactVirtualized__Table__row {
-    @media (max-width: 768px) {
-      width: 100% !important;
-      max-width: 100% !important;
-    }
-  }
-  .ReactVirtualized__Table__row {
-    color: white;
-    padding: 1.25rem;
-    font-family: 'Roboto', sans-serif;
-    background: transparent;
-    &:nth-child(2n) {
-      background: #042d41;
-      border-top: 1px solid #0c3549;
-      border-bottom: 1px solid #0c3549;
-    }
-  }
-
-  .icon {
-    margin-right: 5px;
-  }
-
-  .address {
-    font-family: monospace;
-    font-style: normal;
-    font-weight: normal;
-    padding: 0.2rem 0.5rem;
-    font-size: 1.1em;
-    background: #073d57;
-    border-radius: 0.2rem;
-    display: block;
-    width: 70%;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-
-  .DateRangePickerInput {
-    width: 300px;
-    height: 40px;
-    padding-left: 1px;
-    border-radius: 0.4rem !important;
-    border: 1px solid #c8cdd2;
-    .DateInput,
-    .DateInput_input {
-      background: transparent;
-    }
-  }
-
-  .pagination {
-    justify-content: center;
-    padding: 1rem 0;
-    font-size: 0.9rem;
-    background: #042d41;
-    border-top: 1px solid #0c3549;
-    .pageNumber {
-      padding: 0.2rem 0.5rem;
-      font-size: 1.1em;
-      border-radius: 0.2rem;
-      color: white !important;
-      &.active {
-        background: #107491;
-      }
-    }
-  }
-
-  .Buy {
-    color: #f9787a;
-  }
-
-  .sell {
-    color: #72cc98;
-  }
-}
