--- src/components/bonds/exchange/Exchange.scss
@@ -1,113 +1,0 @@
-@import '../../../assets/styles/colors';
-@import '../../../assets/styles/fonts';
-
-$border-radius: 0.5rem;
-
-.BuySellForm_wrapper {
-  width: 100%;
-}
-
-input {
-  border: 1px solid #c8cdd2;
-  border-radius: $border-radius;
-  background-color: #fff;
-  display: block;
-  box-sizing: border-box;
-  width: 100%;
-  margin-bottom: 0.625rem;
-  font-family: $form-label-font;
-  font-size: $form-label-font-size;
-  padding: 0.375rem 0.625rem;
-  height: 2.5rem;
-  overflow: visible;
-}
-
-select {
-  border: 1px solid #c8cdd2;
-  border-radius: $border-radius;
-  background-color: #fff;
-  margin-bottom: 0.625rem;
-  width: 100%;
-  padding: 0.313rem 2.188rem 0.313rem 0.313rem;
-  font-family: $form-label-font;
-  font-size: $form-label-font-size;
-  height: 2.5rem;
-  text-transform: none;
-}
-
-input,
-select {
-  -webkit-transition: all 0.3s;
-  transition: all 0.3s;
-}
-
-.label {
-  padding-top: 1rem;
-  color: #444;
-  &:first-of-type {
-    border: 0;
-  }
-}
-
-@mixin subtitle {
-  font-family: $form-subtitle-label-font;
-  font-size: $form-subtitle-label-font-size;
-}
-
-.label_subtitle {
-  @include subtitle;
-  color: #828282;
-  text-align: right;
-}
-
-.label_subtitle__bold {
-  @include subtitle;
-  color: #333333;
-  font-weight: bold;
-}
-
-.error {
-  @include subtitle;
-  color: firebrick;
-  text-align: right;
-}
-
-.button {
-  width: 15rem;
-  border-radius: 0.3rem;
-  //letter-spacing: 0.5rem;
-  text-transform: uppercase;
-  border: none;
-  cursor: pointer;
-  font-family: $form-submit-button-font;
-  font-size: $form-label-font-size;
-  -webkit-appearance: none;
-  line-height: 1;
-  //display: inline-block;
-  transition: all 0.2s ease-out 0s;
-}
-
-.button_buy {
-  margin: 2rem 0px 1rem;
-  color: white;
-}
-
-.button_buy_quote {
-  border: 0;
-  background: #2f6f93;
-  background: linear-gradient(180deg, #2f6f93 0%, #3d8cb2 100%);
-  margin-left: auto;
-}
-
-.button_buy_confirm {
-  background: #219653;
-  background: linear-gradient(180deg, #219653 0%, #27ae60 100%);
-  margin-left: auto;
-}
-
-.button_buy_back {
-  border: 1px solid #999;
-  background: transparent;
-  color: #555;
-  margin-right: 3rem;
-}
