--- src/components/bonds/overview/events/Events.tsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-
-export class Events extends Component {
-  render(): JSX.Element {
-    return (
-      <div className="BondsWrapper_panel__chrome">
-        <div className="BondsWrapper_panel__content">
-          <b>Events</b>
-        </div>
-      </div>
-    )
-  }
-}
