+++ src/modules/BondEvents/screens/Events.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export const Events: React.FunctionComponent<any> = () => {
+  return (
+    <div className="BondsWrapper_panel__chrome">
+      <div className="BondsWrapper_panel__content">
+        <b>Events</b>
+      </div>
+    </div>
+  )
+}
