+++ src/modules/account/account.actions.ts
@@ -48,9 +48,14 @@ export const getAccount = (address: string) => (
   })
 }
 
-export const getOrders = (address: string) => (
+export const getOrders = () => (
   dispatch: Dispatch,
+  getState: () => RootState,
 ): GetOrdersAction => {
+  const {
+    account: { address },
+  } = getState()
+
   const config = {
     transformResponse: [
       (response: string): any => {
